requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
typer>=0.9.0,<1.0.0
streamlit>=1.33.0
pdfplumber>=0.10.0
python-docx>=0.8.11
pytest>=7.4.0
